Reimplement modular integer calculator to avoid overflow

+++ src/cn/ancono/math/numberTheory/ZModPCalculator.java
@@ -11,7 +11,7 @@ import cn.ancono.math.MathCalculator;
 public interface ZModPCalculator<T> extends MathCalculator<T> {
 
 
-    int getP();
+    long getP();
 
     @Override
     default long getCharacteristic() {
